message inserted_at_ns

+++ xmtp_mls/migrations/2025-03-28-132353_add_sync_cursor/up.sql
@@ -22,7 +22,7 @@ DROP TABLE user_preferences_old;
 
 -- Add an inserted_at_ns to group_messages
 ALTER TABLE group_messages
-ADD COLUMN inserted_at_ns INTEGER DEFAULT (unixepoch('subsecond') * 1_000_000_000);
+ADD COLUMN inserted_at_ns BIGINT DEFAULT (unixepoch('subsecond') * 1_000_000_000);
 
 -- Set the existing messages inserted_at_ns to sent_at_ns
 UPDATE group_messages
